perf: improve performance across data fetching, bundle size, and rendering
- Remove fetchCache='force-no-store' that disabled all Next.js server caching
- Batch category/tag fetching via WP REST API include parameter (N+1 -> 1 request)
- Parallelize sequential data fetches in blog listing and blog post pages
- Lazy-load framer-motion via dynamic import (header mobile menu)
- Convert footer from full client component to server component with tiny client clock
- Drop unused Inter font weight 200
- Add framer-motion to optimizePackageImports for better tree-shaking
