+++ components/header.tsx
@@ -4,8 +4,10 @@ import Link from "next/link"
 import { usePathname } from "next/navigation"
 import { useState, useEffect, useRef } from "react"
 import { cn } from "@/lib/utils"
-import { Menu, X, ChevronDown, ArrowRight, Layers } from "lucide-react"
-import { motion, AnimatePresence } from "framer-motion"
+import { Menu, ChevronDown, ArrowRight, Layers } from "lucide-react"
+import dynamic from "next/dynamic"
+
+const MobileMenu = dynamic(() => import("./mobile-menu"), { ssr: false })
 
 interface NavigationItem {
   href: string
@@ -196,142 +198,15 @@ export function Header() {
         <Menu className="h-6 w-6 text-zinc-900" />
       </button>
 
-      {/* Mobile Navigation Overlay */}
-      <AnimatePresence>
-        {isOpen && (
-          <div className="md:hidden fixed inset-0 min-h-screen w-screen z-50">
-            {/* Backdrop */}
-            <motion.div
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-              transition={{ duration: 0.2 }}
-              className="absolute inset-0 bg-white/30 backdrop-blur-[20px]"
-              onClick={() => setIsOpen(false)}
-            />
-
-            {/* Full Screen Menu */}
-            <motion.div
-              initial={{ x: '-100%' }}
-              animate={{ x: 0 }}
-              exit={{ x: '-100%' }}
-              transition={{ type: "spring", bounce: 0, duration: 0.4 }}
-              className="absolute inset-0 glass-panel"
-            >
-              {/* Close Button */}
-              <div className="absolute top-6 right-6 z-50">
-                <button
-                  onClick={() => setIsOpen(false)}
-                  aria-label="Close menu"
-                  className="p-2 hover:bg-zinc-100 rounded-lg transition-colors"
-                >
-                  <X className="h-6 w-6 text-zinc-900" />
-                </button>
-              </div>
-
-              {/* Navigation Links */}
-              <div className="flex min-h-screen items-center justify-center pt-16 px-6">
-                <nav className="flex flex-col items-center justify-center space-y-6 w-full">
-                  {/* Home link */}
-                  <Link
-                    href="/"
-                    onClick={() => setIsOpen(false)}
-                    className={cn(
-                      "text-3xl font-light tracking-tight transition-colors hover:text-orange-500",
-                      pathname === '/' ? "text-orange-500" : "text-zinc-900"
-                    )}
-                  >
-                    Home
-                  </Link>
-                  
-                  {navigationItems.map((item) => (
-                    <div key={item.href} className="text-center">
-                      {item.children ? (
-                        <>
-                          <button
-                            onClick={() => toggleSubmenu(item.href)}
-                            className={cn(
-                              "text-3xl font-light tracking-tight transition-colors hover:text-orange-500 inline-flex items-center gap-2",
-                              pathname?.startsWith(item.href)
-                                ? "text-orange-500"
-                                : "text-zinc-900"
-                            )}
-                          >
-                            {item.label}
-                            <ChevronDown 
-                              className={cn(
-                                "w-6 h-6 transition-transform",
-                                expandedItems[item.href] && "rotate-180"
-                              )} 
-                            />
-                          </button>
-                          
-                          <AnimatePresence>
-                            {expandedItems[item.href] && (
-                              <motion.div
-                                initial={{ height: 0, opacity: 0 }}
-                                animate={{ height: 'auto', opacity: 1 }}
-                                exit={{ height: 0, opacity: 0 }}
-                                transition={{ duration: 0.3 }}
-                                className="mt-4 space-y-3 overflow-hidden"
-                              >
-                                {item.children.map((child) => (
-                                  <Link
-                                    key={child.href}
-                                    href={child.href}
-                                    onClick={() => setIsOpen(false)}
-                                    className={cn(
-                                      "block text-xl font-light transition-colors hover:text-orange-500",
-                                      pathname === child.href
-                                        ? "text-orange-500"
-                                        : "text-zinc-500"
-                                    )}
-                                  >
-                                    {child.label}
-                                  </Link>
-                                ))}
-                              </motion.div>
-                            )}
-                          </AnimatePresence>
-                        </>
-                      ) : (
-                        <Link
-                          href={item.href}
-                          onClick={() => setIsOpen(false)}
-                          className={cn(
-                            "text-3xl font-light tracking-tight transition-colors hover:text-orange-500",
-                            pathname === item.href
-                              ? "text-orange-500"
-                              : "text-zinc-900"
-                          )}
-                        >
-                          {item.label}
-                        </Link>
-                      )}
-                    </div>
-                  ))}
-
-                  {/* Hire Me CTA in mobile */}
-                  <Link
-                    href="https://www.upwork.com/freelancers/~0117c4a4c888d9e9fe"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    onClick={() => setIsOpen(false)}
-                    className="mt-8 rounded-full py-4 px-10 text-base font-semibold flex items-center gap-3 text-white"
-                    style={{
-                      background: 'linear-gradient(135deg, #14a800 0%, #0d7a00 100%)',
-                      boxShadow: '0 10px 30px rgba(20, 168, 0, 0.4)',
-                    }}
-                  >
-                    Hire Me on Upwork
-                    <ArrowRight className="w-5 h-5" />
-                  </Link>
-                </nav>
-              </div>
-            </motion.div>
-          </div>
-        )}
-      </AnimatePresence>
+      {/* Mobile Navigation Overlay (lazy-loaded with framer-motion) */}
+      <MobileMenu
+        isOpen={isOpen}
+        onClose={() => setIsOpen(false)}
+        pathname={pathname}
+        navigationItems={navigationItems}
+        expandedItems={expandedItems}
+        toggleSubmenu={toggleSubmenu}
+      />
     </header>
   )
 }
